[refactor/#40] FSD 레이어 분리 및 위젯화 (#40)
