+++ src/entities/recipe/index.ts
@@ -1,3 +1,5 @@
 // Public API for recipe entity model and UI.
+export type { CookingStep, Ingredient, Recipe } from "./model/recipe.types";
 export type { RecipeCardData, RecipeCardVariant } from "./model/types";
+export { useRecipeDetail } from "./model/useRecipeDetail";
 export { BannerFoodCard, DefaultFoodCard } from "./ui";
