+++ src/features/search-recipe/index.ts
@@ -1,0 +1,2 @@
+// Public API for search-recipe feature.
+export { useSearchRecipe } from "./hooks/useSearchRecipe";
