+++ src/widgets/RecipeDetail/index.ts
@@ -1,0 +1,5 @@
+// Public API for RecipeDetail widgets.
+export { RecipeHero } from "./RecipeHero";
+export { RecipeIngredients } from "./RecipeIngredients";
+export { RecipeMeta } from "./RecipeMeta";
+export { RecipeSteps } from "./RecipeSteps";
